Fix duplicate codes
